requirements.txt
crewai==0.30.11
crewai-tools==0.2.6
langchain==0.1.20           
httpx<0.28
openai==1.30.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
ipython==8.24.0
ipykernel==6.29.4
jupyter_client==8.6.2
jupyter_core==5.7.2
tiktoken==0.5.2
yfinance==0.2.40
